Rename files, and remove outdated files.
