+++ cloudmesh/compute/aws/AwsFlavor.py
